rename context to userData for fgr vorabcheck
