+++ app/domains/fluggastrechte/vorabcheck/__test__/guards.test.ts
@@ -1,6 +1,6 @@
 import { Result } from "true-myth";
 import { calculateDistanceBetweenAirportsInKilometers } from "~/domains/fluggastrechte/services/airports/calculateDistanceBetweenAirports";
-import type { FluggastrechtVorabcheckContext } from "~/domains/fluggastrechte/vorabcheck/context";
+import type { FluggastrechtVorabcheckUserData } from "~/domains/fluggastrechte/vorabcheck/userData";
 import { guards } from "~/domains/fluggastrechte/vorabcheck/guards";
 
 vi.mock(
@@ -13,7 +13,7 @@ const mockedCalculateDistanceBetweenAirportsInKilometers = vi.mocked(
 describe("fluggastrechteGuard", () => {
   describe("areAirportsOutsideEU", () => {
     it("if the departure airport is in EU country, it should return false", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BER",
         endAirport: "GRU",
       };
@@ -24,7 +24,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("if the destination airport is in EU country, it should return false", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "GRU",
         endAirport: "BER",
       };
@@ -35,7 +35,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("if the both airports are outside EU, it should return true", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "GRU",
       };
@@ -46,7 +46,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("if the departure airport does not exist, it should return false", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BLABLA",
         endAirport: "BER",
       };
@@ -57,7 +57,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("if the destination airport does not exist, it should return false", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BER",
         endAirport: "BLABLA",
       };
@@ -68,7 +68,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("if the departure airport is undefined, it should return false", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         endAirport: "BER",
       };
 
@@ -78,7 +78,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("if the destination airport is undefined, it should return true", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BER",
       };
 
@@ -94,7 +94,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("in case is not possible to calculate the airports, it should return true", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BER",
       };
 
@@ -108,7 +108,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("in case is possible to calculate the airports, it should return false", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BER",
       };
 
@@ -124,7 +124,7 @@ describe("fluggastrechteGuard", () => {
 
   describe("isNonGermanAirportsAndIsNotClaimableInEUWithOtherAirline", () => {
     it("should return false given a german departure airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BER",
         endAirport: "CDG",
         fluggesellschaft: "sonstiges",
@@ -139,7 +139,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a german destination airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "CDG",
         endAirport: "BER",
         fluggesellschaft: "sonstiges",
@@ -154,7 +154,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a departure EU airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "CDG",
         endAirport: "JFK",
         fluggesellschaft: "sonstiges",
@@ -169,7 +169,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a non departure EU airport, destination EU airport and EU airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "CDG",
         fluggesellschaft: "LH",
@@ -184,7 +184,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return true given a non departure EU airport, destination EU airport and sonstiges airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "CDG",
         fluggesellschaft: "sonstiges",
@@ -201,7 +201,7 @@ describe("fluggastrechteGuard", () => {
 
   describe("isNonGermanAirportsAndIsNotClaimableInEU", () => {
     it("should return false given a german departure airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BER",
         endAirport: "CDG",
         fluggesellschaft: "LH",
@@ -215,7 +215,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a german destination airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "CDG",
         endAirport: "BER",
         fluggesellschaft: "LH",
@@ -229,7 +229,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a departure EU airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "CDG",
         endAirport: "JFK",
         fluggesellschaft: "LH",
@@ -243,7 +243,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a non departure EU airport, destination EU airport and EU airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "CDG",
         fluggesellschaft: "LH",
@@ -257,7 +257,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return true given a non departure EU airport, destination EU airport and non EU airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "CDG",
         fluggesellschaft: "DL",
@@ -273,7 +273,7 @@ describe("fluggastrechteGuard", () => {
 
   describe("isErfolgEU", () => {
     it("should return false given gericht as yes", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "CDG",
         endAirport: "AMS",
         fluggesellschaft: "LH",
@@ -288,7 +288,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a german departure airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "BER",
         endAirport: "AMS",
         fluggesellschaft: "LH",
@@ -303,7 +303,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a german destination airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "CDG",
         endAirport: "BER",
         fluggesellschaft: "LH",
@@ -318,7 +318,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return true given an EU airports ", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "CDG",
         endAirport: "AMS",
         fluggesellschaft: "LH",
@@ -333,7 +333,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given an EU destination airport, non EU departure airport and airline ", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "AMS",
         fluggesellschaft: "DL",
@@ -348,7 +348,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return true given an EU destination airport and airline, non EU departure airport ", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "AMS",
         fluggesellschaft: "LH",
@@ -363,7 +363,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given non EU departure and destination airports", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "GRU",
         fluggesellschaft: "LH",
@@ -380,7 +380,7 @@ describe("fluggastrechteGuard", () => {
 
   describe("isGermanEndAirportsAndIsNotClaimable", () => {
     it("should return false given a departure german airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "DRS",
         endAirport: "BER",
         fluggesellschaft: "DL",
@@ -394,7 +394,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a destination german airport, non eu departure and eu airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "BER",
         fluggesellschaft: "LH",
@@ -408,7 +408,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a destination german airport, non eu departure and sonstiges airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "BER",
         fluggesellschaft: "sonstiges",
@@ -422,7 +422,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a destination german airport, eu airline and non eu airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "AMS",
         endAirport: "DRS",
         fluggesellschaft: "DL",
@@ -436,7 +436,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a destination german airport, eu departure and eu airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "AMS",
         endAirport: "DRS",
         fluggesellschaft: "LH",
@@ -450,7 +450,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a destination german airport, eu departure and sonstiges airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "AMS",
         endAirport: "DRS",
         fluggesellschaft: "sonstiges",
@@ -464,7 +464,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return true given a destination german airport, non eu departure and non eu airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "DRS",
         fluggesellschaft: "DL",
@@ -480,7 +480,7 @@ describe("fluggastrechteGuard", () => {
 
   describe("isGermanEndAirportsAndOtherAirline", () => {
     it("should return false given both non german airports", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "AMS",
         endAirport: "CDG",
         fluggesellschaft: "sonstiges",
@@ -494,7 +494,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a departure german airport", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "DRS",
         endAirport: "DRS",
         fluggesellschaft: "sonstiges",
@@ -508,7 +508,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given an eu airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "AMS",
         endAirport: "DRS",
         fluggesellschaft: "LH",
@@ -522,7 +522,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a destination german airport, an eu departure airport and sonstiges airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "AMS",
         endAirport: "DRS",
         fluggesellschaft: "sonstiges",
@@ -536,7 +536,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return false given a destination german airport, a non eu departure airport and an eu airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "DRS",
         fluggesellschaft: "LH",
@@ -550,7 +550,7 @@ describe("fluggastrechteGuard", () => {
     });
 
     it("should return true given a destination german airport, a non eu departure airport and sonstiges airline", () => {
-      const context: FluggastrechtVorabcheckContext = {
+      const context: FluggastrechtVorabcheckUserData = {
         startAirport: "JFK",
         endAirport: "DRS",
         fluggesellschaft: "sonstiges",
