+++ app/domains/fluggastrechte/vorabcheck/__test__/testcasesNichtBefoerderungAbbruch.ts
@@ -1,6 +1,6 @@
 import { createMachine } from "xstate";
 import type { TestCases } from "~/domains/__test__/TestCases";
-import type { FluggastrechtVorabcheckContext } from "~/domains/fluggastrechte/vorabcheck/context";
+import type { FluggastrechtVorabcheckUserData } from "~/domains/fluggastrechte/vorabcheck/userData";
 import { guards } from "~/domains/fluggastrechte/vorabcheck/guards";
 import type { FlowStateMachine } from "~/services/flow/server/types";
 import { fluggastrechteVorabcheckXstateConfig } from "../xstateConfig";
@@ -189,7 +189,7 @@ const cases = [
       "/ergebnis/abtretung-abbruch",
     ],
   ],
-] as const satisfies TestCases<FluggastrechtVorabcheckContext>;
+] as const satisfies TestCases<FluggastrechtVorabcheckUserData>;
 
 export const testCasesFluggastrechteNichtBefoerderungAbbruch = {
   machine,
