+++ app/domains/fluggastrechte/vorabcheck/services/__test__/isErfolgAnalog.test.ts
@@ -1,4 +1,4 @@
-import { type FluggastrechtVorabcheckContext } from "../../context";
+import { type FluggastrechtVorabcheckUserData } from "../../userData";
 import { isErfolgAnalog } from "../isErfolgAnalog";
 
 const EU_AIRLINE = "LH";
@@ -7,7 +7,7 @@ const SONSTIGES_AIRLINE = "sonstiges";
 
 describe("isErfolgAnalog", () => {
   it("should return false given gericht as yes", () => {
-    const context: FluggastrechtVorabcheckContext = {
+    const context: FluggastrechtVorabcheckUserData = {
       startAirport: "AMS",
       endAirport: "DRS",
       fluggesellschaft: "LH",
@@ -22,7 +22,7 @@ describe("isErfolgAnalog", () => {
   describe("departure outside EU and only destination in DE", () => {
     describe("destination without partner court", () => {
       it("should return true given a departure outside EU and eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "JFK",
           endAirport: "DRS",
           fluggesellschaft: EU_AIRLINE,
@@ -35,7 +35,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return false given a departure outside EU and non eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "JFK",
           endAirport: "DRS",
           fluggesellschaft: NON_EU_AIRLINE,
@@ -48,7 +48,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return false given a departure outside EU and airline sonstiges", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "JFK",
           endAirport: "DRS",
           fluggesellschaft: SONSTIGES_AIRLINE,
@@ -63,7 +63,7 @@ describe("isErfolgAnalog", () => {
 
     describe("destination with partner court", () => {
       it("should return false given a departure outside EU and eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "JFK",
           endAirport: "BER",
           fluggesellschaft: EU_AIRLINE,
@@ -80,7 +80,7 @@ describe("isErfolgAnalog", () => {
   describe("only departure DE or both in DE", () => {
     describe("without partner court", () => {
       it("should return true given eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "DRS",
           endAirport: "ERF",
           fluggesellschaft: EU_AIRLINE,
@@ -93,7 +93,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return true given non eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "DRS",
           endAirport: "ERF",
           fluggesellschaft: NON_EU_AIRLINE,
@@ -106,7 +106,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return true given sonstiges airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "DRS",
           endAirport: "ERF",
           fluggesellschaft: SONSTIGES_AIRLINE,
@@ -121,7 +121,7 @@ describe("isErfolgAnalog", () => {
 
     describe("with partner court", () => {
       it("should return true departure airport with partner court and sonstiges airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "BER",
           endAirport: "ERF",
           fluggesellschaft: SONSTIGES_AIRLINE,
@@ -134,7 +134,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return false departure airport with partner court and eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "BER",
           endAirport: "ERF",
           fluggesellschaft: EU_AIRLINE,
@@ -147,7 +147,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return false departure airport with partner court and non eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "BER",
           endAirport: "ERF",
           fluggesellschaft: NON_EU_AIRLINE,
@@ -160,7 +160,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return true destination airport with partner court and sonstiges airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "ERF",
           endAirport: "BER",
           fluggesellschaft: SONSTIGES_AIRLINE,
@@ -173,7 +173,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return false destination airport with partner court and eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "ERF",
           endAirport: "BER",
           fluggesellschaft: EU_AIRLINE,
@@ -186,7 +186,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return false destination airport with partner court and non eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "ERF",
           endAirport: "BER",
           fluggesellschaft: NON_EU_AIRLINE,
@@ -203,7 +203,7 @@ describe("isErfolgAnalog", () => {
   describe("departure in EU and destination in DE", () => {
     describe("destination without partner court", () => {
       it("should return true given eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "AMS",
           endAirport: "ERF",
           fluggesellschaft: EU_AIRLINE,
@@ -216,7 +216,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return true given non eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "AMS",
           endAirport: "ERF",
           fluggesellschaft: NON_EU_AIRLINE,
@@ -229,7 +229,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return true given non sonstiges airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "AMS",
           endAirport: "ERF",
           fluggesellschaft: SONSTIGES_AIRLINE,
@@ -244,7 +244,7 @@ describe("isErfolgAnalog", () => {
 
     describe("destination with partner court", () => {
       it("should return false given eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "AMS",
           endAirport: "BER",
           fluggesellschaft: EU_AIRLINE,
@@ -257,7 +257,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return false given non eu airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "AMS",
           endAirport: "BER",
           fluggesellschaft: NON_EU_AIRLINE,
@@ -270,7 +270,7 @@ describe("isErfolgAnalog", () => {
       });
 
       it("should return false given sonstiges airline", () => {
-        const context: FluggastrechtVorabcheckContext = {
+        const context: FluggastrechtVorabcheckUserData = {
           startAirport: "AMS",
           endAirport: "BER",
           fluggesellschaft: SONSTIGES_AIRLINE,
